+++ django_api/django_api/apps/core/tests/test_views.py
@@ -1,10 +1,10 @@
 # from django.urls import reverse
 
-# from rest_framework import status
+# from rest_framework imports status
 
-# from core.models import ResponsePlan
-# from core.models import Location, Workspace
-# from .base import BaseAPITestCase
+# from core.models imports ResponsePlan
+# from core.models imports Location, Workspace
+# from .base imports BaseAPITestCase
 
 
 # class TestWorkspaceListAPIView(BaseAPITestCase):
