+++ django_api/django_api/settings/__init__.py
@@ -1,6 +1,6 @@
 from importlib import import_module
 
-# import defaults
+# imports defaults
 from .base import *
 
 # This will make sure the app is always imported when
@@ -13,7 +13,7 @@ overrides = import_module('django_api.settings.{}'.format(ENV))
 
 # apply imported overrides
 for attr in dir(overrides):
-    # we only want to import settings (which have to be variables in ALLCAPS)
+    # we only want to imports settings (which have to be variables in ALLCAPS)
     if attr.isupper():
         # update our scope with the imported variables. We use globals() instead of locals()
         # because locals() is readonly and it returns a copy of itself upon
