+++ polymer/index.html
@@ -95,7 +95,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
         });
       }
 
-      // set app shell import path based on baseUrl
+      // set app shell imports path based on baseUrl
       var appShellImport = document.createElement('link');
       appShellImport.rel = 'import';
       appShellImport.href = baseUrl + 'src/app-shell.html';
